Hopefully fix all remaining quantizer issues in Mario Kart Wii:   * must use a truncating float-to-int conversion, for example.   * introduce optimized variants of the single value psq_st operation (JIT only).   * fix bug in SafeWriteRegToReg when swap = false
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@4861 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/PowerPC/Interpreter/Interpreter_LoadStorePaired.cpp
@@ -74,9 +74,9 @@ inline T CLAMP(T a, T bottom, T top) {
 }
 
 void Helper_Quantize(const u32 _Addr, const double _fValue, 
-							  const EQuantizeType _quantizeType, const unsigned int _uScale)
+					 const EQuantizeType _quantizeType, const unsigned int _uScale)
 {
-	switch(_quantizeType) 
+	switch (_quantizeType) 
 	{
 	case QUANTIZE_FLOAT:		
 		Memory::Write_U32( ConvertToSingleFTZ( *(u64*)&_fValue ), _Addr );
@@ -222,7 +222,7 @@ void psq_st(UGeckoInstruction _inst)
 	}
 	else
 	{
-		Helper_Quantize( EA,   (float)rPS0(_inst.RS), stType, stScale );
+		Helper_Quantize( EA,   rPS0(_inst.RS), stType, stScale );
 	}
 }
